% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_compass.R
\name{plot_compass}
\alias{plot_compass}
\title{Political compass}
\usage{
plot_compass(data)
}
\arguments{
\item{data}{A table of values to be plot. Input requires two columns for
'Economic' and 'Social' values.}
}
\value{
None
}
\description{
Create a scatter plot of "The Political Compass" results.
}
\references{
\url{https://www.politicalcompass.org}
}
\author{
Fred Jaya
}
